requirements.txt
opencv-python-headless>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
Flask>=3.0.0
firebase-admin>=6.5.0
python-dotenv>=1.0.1
tensorflow>=2.15.1
gunicorn>=21.2.0
matplotlib>=3.7.0
scipy>=1.11.0